requirements.txt
Django==4.2.7
pymongo==4.6.0
grpcio==1.59.2
grpcio-tools==1.59.2
mysqlclient==2.2.0
python-dotenv==1.0.0
dnspython==2.4.2